requirements.txt
flask==2.3.3
torch>=2.0.0
transformers>=4.39.0
pillow>=10.0.0
accelerate>=0.27.0
gunicorn==21.2.0